Pybind BS caused more overhead than was saved :(

+++ scripts/bit_tools.py
@@ -1,19 +1,18 @@
-from uint128 import UInt128
 
 # import itertools
 def as_mask(spin_indices: list) -> int:
-    mask = UInt128(0)
+    mask = 0
     for j in spin_indices:
-        mask |= UInt128(1) << j
+        mask |= 1 << j
     return mask
 
 
 def make_state(spin_indices: list, state: int):
     # state = n-bit integer
-    mask = UInt128(0)
+    mask = 0
     i = 0
     for spin_idx in spin_indices:
-        mask |= ((state & (UInt128(1) << i)) >> i) << spin_idx
+        mask |= ((state & (1 << i)) >> i) << spin_idx
         i += 1
     return mask
 
@@ -27,10 +26,10 @@ def make_state(spin_indices: list, state: int):
 #     return arr.reshape(-1, la)
 
 
-def bitperm(perm: list,  x: UInt128):
+def bitperm(perm: list,  x: int):
     # does the specified permutation on the bits of x
-    y = UInt128(0)
+    y = 0
     for i in range(len(perm)):
-        y |= ((x & (UInt128(1)<<i)) >> i) << perm[i]
+        y |= ((x & (1<<i)) >> i) << perm[i]
 
     return y
